Adapt our handlers / plugins to FC 5
