+++ src/fullcalendar/rendering/eventContent.js
