+++ tests/javascript/unit/fullcalendar/interaction/navLinkWeekClick.test.js
@@ -21,7 +21,7 @@
  */
 import navLinkWeekClick from '../../../../../src/fullcalendar/interaction/navLinkWeekClick.js'
 import { getYYYYMMDDFromDate } from '../../../../../src/utils/date.js'
-jest.mock('../../../../src/utils/date.js')
+jest.mock('../../../../../src/utils/date.js')
 
 describe('fullcalendar/eventClick test suite', () => {
